Move prices to the type BigInteger

+++ src/main/java/com/iba/project/entity/plan/abstraction/interfaces/PlanInterface.java
@@ -1,15 +1,17 @@
 package com.iba.project.entity.plan.abstraction.interfaces;
 
+import java.math.BigDecimal;
+
 public interface PlanInterface {
     public String getCellularName();
 
-    public int getPrice();
+    public BigDecimal getPrice();
 
     public String getName();
 
     public void setCellularName(String name);
 
-    public void setPrice(int price);
+    public void setPrice(BigDecimal price);
 
     public void setName(String name);
