+++ src/main/java/com/iba/project/entity/plan/abstraction/interfaces/PlanWithInternetInterface.java
@@ -1,7 +1,9 @@
 package com.iba.project.entity.plan.abstraction.interfaces;
 
+import java.math.BigDecimal;
+
 public interface PlanWithInternetInterface
 {
-    int getPriceOfInternet();
-    void setPriceOfInternet(int priceForInternet);
+    BigDecimal getPriceOfInternet();
+    void setPriceOfInternet(BigDecimal priceForInternet);
 }
